adds personalized data

+++ service/Controllers/BluRaysController.cs
@@ -22,8 +22,32 @@ public class BluRaysController : ControllerBase
   [HttpGet]
   public async Task<ActionResult<IEnumerable<BluRay>>> GetBluRays()
   {
-    //GetBluRays method, add parameter of string searchString to above method declaration- entity framework filtering
-    return await _context.BluRays.ToListAsync();
+    var request = Request;
+    var headers = request.Headers;
+
+    if (headers.ContainsKey("owner"))
+    {
+      //GetBluRays method, add parameter of string searchString to above method declaration- entity framework filtering
+      var foundBluRays = await _context.BluRays.ToListAsync();
+      List<BluRay> returnedBluRays = new List<BluRay> { };
+
+      for (int i = 0; i < foundBluRays.Count; i++)
+      {
+        string dbOwner = foundBluRays[i].Owner;
+        string rqOwner = headers["owner"];
+        if (Equals(dbOwner, rqOwner))
+        {
+          returnedBluRays.Add(foundBluRays[i]);
+        }
+      }
+
+      return returnedBluRays;
+    }
+    else
+    {
+      return NotFound();
+    }
+
   }
 
   // GET: api/blurays/5
@@ -31,53 +55,87 @@ public class BluRaysController : ControllerBase
   public async Task<ActionResult<BluRay>> GetBluRay(int id)
   {
     // Use below as example to query database, make separate class as a service (to be called by the controller) that is the only one that calls the database
-    var bluRay = await _context.BluRays.FindAsync(id);
 
-    if (bluRay == null)
+    var request = Request;
+    var headers = request.Headers;
+
+    if (headers.ContainsKey("owner"))
     {
-      return NotFound();
+      var bluRay = await _context.BluRays.FindAsync(id);
+
+      string dbOwner = bluRay.Owner;
+      string rqOwner = headers["owner"];
+
+      if (bluRay == null || !Equals(dbOwner, rqOwner))
+      {
+        return NotFound();
+      }
+      return bluRay;
     }
 
-    return bluRay;
+    return NotFound();
+
   }
 
   // POST: api/blurays
   [HttpPost]
   public async Task<ActionResult<BluRay>> PostBluRay(BluRay bluRay)
   {
-    _context.BluRays.Add(bluRay);
-    await _context.SaveChangesAsync();
+    var request = Request;
+    var headers = request.Headers;
 
-    return CreatedAtAction(nameof(GetBluRay), new { id = bluRay.Id }, bluRay);
+    if (headers.ContainsKey("owner"))
+    {
+      _context.BluRays.Add(bluRay);
+      await _context.SaveChangesAsync();
+
+      return CreatedAtAction(nameof(GetBluRay), new { id = bluRay.Id }, bluRay);
+    }
+    return BadRequest();
   }
 
   // PUT: api/blurays/5
   [HttpPut("{id}")]
   public async Task<IActionResult> PutBluRay(int id, BluRay bluRay)
   {
-    if (id != bluRay.Id)
-    {
-      return BadRequest();
-    }
+    var request = Request;
+    var headers = request.Headers;
 
-    _context.Entry(bluRay).State = EntityState.Modified;
-
-    try
+    if (headers.ContainsKey("owner"))
     {
-      await _context.SaveChangesAsync();
-    }
-    catch (DbUpdateConcurrencyException)
-    {
-      if (!BluRayExists(id))
+      if (id != bluRay.Id)
       {
-        return NotFound();
+        return BadRequest();
+      }
+
+      string dbOwner = bluRay.Owner;
+      string rqOwner = headers["owner"];
+
+      if (Equals(dbOwner, rqOwner))
+      {
+        _context.Entry(bluRay).State = EntityState.Modified;
+
+        try
+        {
+          await _context.SaveChangesAsync();
+        }
+        catch (DbUpdateConcurrencyException)
+        {
+          if (!BluRayExists(id))
+          {
+            return NotFound();
+          }
+          else
+          {
+            throw;
+          }
+        }
       }
       else
       {
-        throw;
+        return BadRequest();
       }
     }
-
     return NoContent();
   }
 
@@ -85,38 +143,35 @@ public class BluRaysController : ControllerBase
   [HttpDelete("{id}")]
   public async Task<IActionResult> DeleteBluRay(int id)
   {
-    var bluRay = await _context.BluRays.FindAsync(id);
-    if (bluRay == null)
+    var request = Request;
+    var headers = request.Headers;
+
+    if (headers.ContainsKey("owner"))
+    {
+      var bluRay = await _context.BluRays.FindAsync(id);
+
+      string dbOwner = bluRay.Owner;
+      string rqOwner = headers["owner"];
+
+      if (bluRay == null ||!Equals(dbOwner, rqOwner))
+      {
+        return NotFound();
+      }
+
+      _context.BluRays.Remove(bluRay);
+      await _context.SaveChangesAsync();
+
+      return NoContent();
+    }
+    else
     {
       return NotFound();
     }
-
-    _context.BluRays.Remove(bluRay);
-    await _context.SaveChangesAsync();
-
-    return NoContent();
   }
 
   private bool BluRayExists(int id)
   {
     return _context.BluRays.Any(e => e.Id == id);
-  }
-
-  // dummy method to test the connection
-  [HttpGet("hello")]
-  public string Test()
-  {
-    return "Hello World!";
-  }
-
-  [HttpGet("private")]
-  [Authorize]
-  public IActionResult Private()
-  {
-    return Ok(new
-    {
-      Message = "Hello from a private endpoint! You need to be authenticated to see this."
-    });
-  }
+  }  
 }
 // Dummy comment
